% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uadfuncfile.R
\name{uadfunc}
\alias{uadfunc}
\title{This is a function}
\usage{
uadfunc(x = 5)
}
\arguments{
\item{x}{It's a parameter that defaults to 5}
}
\description{
This function is for xxx
}
\examples{
uadfunc(3)
}
\keyword{kws}
