\name{UpAndDownPlots}
\alias{UpAndDownPlots-package}
\docType{package}
\title{
UpAndDownPlots: display both percentage and absolute changes
}

\description{
UpAndDown plots display percentage changes by height and absolute changes by area for up to three nested or non-nested levels.  They can visualise changes in indices, showing how the changes for sectors or for individual components contribute to the overall change.
}

\details{
An UpAndDown plot shows both percentage and absolute changes for a variable recorded at two separate time points.
}

\author{
Antony Unwin unwin@math.uni-augsburg.de

}


